adapt TwitterPlatform to Xbase integration
Actions that require a primitive type are now typed. The platform plugin contains a MANIFEST.MF file that is used in eclipse-based environment to resolve external types (typically JVM types).
To manipulate the .platform file in eclipse: import the project as an eclipse project (using the .project file in platform/) instead of a maven project, this way all the runtime dependencies will be imported by default.